requirements.txt
requests
pandas
tqdm
gspread
python-dotenv